Make itself in Docker

+++ make-sloader-itself.sh
@@ -3,6 +3,12 @@
 cd $(git rev-parse --show-toplevel)
 SLOADER_PATH=$(realpath ./build/sloader)
 
+if [ ! -f ${SLOADER_PATH} ]
+then
+    echo "sloader not found at ${SLOADER_PATH}"
+    exit 1
+fi
+
 rm -rf build2
 mkdir -p build2
 cd build2
